requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio @ file:///C:/miniconda3/conda-bld/anyio_1758622521119/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/miniconda3/conda-bld/argon2-cffi-bindings_1757925007150/work
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
async-lru @ file:///C:/miniconda3/conda-bld/async-lru_1761121591843/work
attrs @ file:///C:/miniconda3/conda-bld/attrs_1762356954445/work
babel @ file:///C:/b/abs_ffzt1bmjth/croot/babel_1737454394148/work
beautifulsoup4 @ file:///C:/b/abs_95vq944bnw/croot/beautifulsoup4-split_1756453107890/work
bleach @ file:///C:/b/abs_925i9psm3u/croot/bleach_1732292896852/work
blinker==1.9.0
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1762950484957/work
cachetools==6.2.2
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135349059/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792265/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761745005832/work
click==8.3.1
colorama @ file:///C:/ci_311/colorama_1676422310965/work
comm @ file:///C:/miniconda3/conda-bld/comm_1763119240032/work
debugpy @ file:///C:/miniconda3/conda-bld/debugpy_1762421657701/work
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341249812/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
executing @ file:///C:/miniconda3/conda-bld/executing_1757061250618/work
fastjsonschema @ file:///C:/miniconda3/conda-bld/python-fastjsonschema_1763718732564/work
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.52.0
h11 @ file:///C:/miniconda3/conda-bld/h11_1761931279881/work
httpcore @ file:///C:/b/abs_25m7_xthp6/croot/httpcore_1748526065845/work
httpx @ file:///C:/miniconda3/conda-bld/httpx_1760447374704/work
idna @ file:///C:/miniconda3/conda-bld/idna_1761912044050/work
ipykernel @ file:///C:/miniconda3/conda-bld/ipykernel_1762946093310/work
ipython @ file:///C:/miniconda3/conda-bld/ipython_1762789323276/work
ipython_pygments_lexers @ file:///C:/b/abs_b66hj1lo19/croot/ipython_pygments_lexers_1744753262043/work
ipywidgets @ file:///C:/miniconda3/conda-bld/ipywidgets_1761560180311/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
json5 @ file:///C:/miniconda3/conda-bld/json5_1763718659894/work
jsonschema @ file:///C:/miniconda3/conda-bld/jsonschema_1762459313135/work
jsonschema-specifications @ file:///C:/miniconda3/conda-bld/jsonschema-specifications_1762788559143/work
jupyter @ file:///C:/b/abs_32aohix0i1/croot/jupyter_1737645821205/work
jupyter-console @ file:///C:/b/abs_82xaa6i2y4/croot/jupyter_console_1680000189372/work
jupyter-events @ file:///C:/b/abs_9cm3qlticu/croot/jupyter_events_1741184612840/work
jupyter-lsp @ file:///C:/b/abs_7171flzdkg/croot/jupyter-lsp-meta_1745827033118/work
jupyter_client @ file:///C:/miniconda3/conda-bld/jupyter_client_1762943100278/work
jupyter_core @ file:///C:/b/abs_7c6fipznw3/croot/jupyter_core_1751991383524/work
jupyter_server @ file:///C:/b/abs_cdpu8und98/croot/jupyter_server_1751996637146/work
jupyter_server_terminals @ file:///C:/b/abs_adjrm9dtns/croot/jupyter_server_terminals_1744706714294/work
jupyterlab @ file:///C:/miniconda3/conda-bld/jupyterlab_1758125109182/work
jupyterlab_pygments @ file:///C:/b/abs_d5alfet8m6/croot/jupyterlab_pygments_1741124274578/work
jupyterlab_server @ file:///C:/miniconda3/conda-bld/jupyterlab_server_1761751445459/work
jupyterlab_widgets @ file:///C:/b/abs_6566by8dag/croot/jupyterlab_widgets_1752066574111/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib-inline @ file:///C:/miniconda3/conda-bld/matplotlib-inline_1762779272071/work
mistune @ file:///C:/b/abs_77yql3poyz/croot/mistune_1741124004410/work
narwhals==2.12.0
nbclient @ file:///C:/b/abs_a09c4t3h8x/croot/nbclient_1741124030330/work
nbconvert @ file:///C:/b/abs_c27_60dzt8/croot/nbconvert-meta_1741191385337/work
nbformat @ file:///C:/b/abs_c2jkw46etm/croot/nbformat_1728050303821/work
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
notebook @ file:///C:/b/abs_ddnsjx06fh/croot/notebook_1756709315423/work
notebook_shim @ file:///C:/b/abs_9ctyfgpncn/croot/notebook-shim_1741707829491/work
numpy==2.3.5
overrides @ file:///C:/miniconda3/conda-bld/overrides_1762349150310/work
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pandas==2.3.3
pandocfilters @ file:///C:/miniconda3/conda-bld/pandocfilters_1756977624481/work
parso @ file:///C:/miniconda3/conda-bld/parso_1762781926029/work
pillow==12.0.0
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356607961/work
prometheus_client @ file:///C:/b/abs_8b175q_ub8/croot/prometheus_client_1744271638821/work
prompt_toolkit @ file:///C:/miniconda3/conda-bld/prompt-toolkit_1761745034319/work
protobuf==6.33.1
psutil @ file:///C:/miniconda3/conda-bld/psutil_1761896536656/work
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067065603/work
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496060165/work
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431425708/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753028158/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.2.1
python-json-logger @ file:///C:/b/abs_0cm_mnox0z/croot/python-json-logger_1734370042436/work
pytz==2025.2
pywin32 @ file:///C:/b/abs_6c9w_vp9zi/croot/pywin32_1754670534373/work
pywinpty @ file:///C:/b/abs_883wh7sts8/croot/pywinpty_1741871674963/work
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465512847/work
pyzmq @ file:///C:/miniconda3/conda-bld/pyzmq_1762375608224/work
qtconsole @ file:///C:/miniconda3/conda-bld/qtconsole_1758748653751/work
QtPy @ file:///C:/miniconda3/conda-bld/qtpy_1757531228141/work
referencing @ file:///C:/miniconda3/conda-bld/referencing_1762536905217/work
requests @ file:///C:/miniconda3/conda-bld/requests_1762359611410/work
rfc3339-validator @ file:///C:/b/abs_ddfmseb_vm/croot/rfc3339-validator_1683077054906/work
rfc3986-validator @ file:///C:/b/abs_6e9azihr8o/croot/rfc3986-validator_1683059049737/work
rpds-py @ file:///C:/miniconda3/conda-bld/rpds-py_1762366642821/work
rsa==4.9.1
Send2Trash @ file:///C:/b/abs_7e73ol18dl/croot/send2trash_1736542724140/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap==5.0.2
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067061065/work
streamlit==1.51.0
tenacity==9.1.2
terminado @ file:///C:/miniconda3/conda-bld/terminado_1757934979771/work
tinycss2 @ file:///C:/b/abs_df38owi5ma/croot/tinycss2_1738337725183/work
toml==0.10.2
tornado @ file:///C:/b/abs_7dzpc171lf/croot/tornado_1748956950306/work
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
typing-inspection==0.4.2
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog==6.0.0
wcwidth @ file:///C:/b/abs_0c8p4c33bp/croot/wcwidth_1750352902378/work
webencodings==0.5.1
websocket-client @ file:///C:/b/abs_5dmnxxoci9/croot/websocket-client_1715878351319/work
websockets==15.0.1
widgetsnbextension @ file:///C:/miniconda3/conda-bld/widgetsnbextension_1761556763484/work
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278197/work
